CI: fix build script

See merge request fontconfig/fontconfig!377

+++ .gitlab-ci/build.sh
@@ -16,9 +16,10 @@ disable=()
 distcheck=0
 enable_install=1
 disable_check=0
+clean_build=1
 cross=0
-buildsys="autotools"
-type="both"
+buildsys="meson"
+type="shared"
 arch=""
 buildopt=()
 SRCDIR=$MyPWD
@@ -27,7 +28,7 @@ export BUILD_ID=${BUILD_ID:-fontconfig-$$}
 export PREFIX=${PREFIX:-$MyPWD/prefix}
 export BUILDDIR=${BUILDDIR:-$MyPWD/build}
 
-while getopts a:cCe:d:hIs:t:X: OPT
+while getopts a:cCe:d:hINs:t:X: OPT
 do
     case $OPT in
         'a') arch=$OPTARG ;;
@@ -36,6 +37,7 @@ do
         'e') enable+=($OPTARG) ;;
         'd') disable+=($OPTARG) ;;
         'I') enable_install=0 ;;
+        'N') clean_build=0 ;;
         's') buildsys=$OPTARG ;;
         't') type=$OPTARG ;;
         'X') backend=$OPTARG ;;
@@ -53,6 +55,18 @@ esac
 
 env
 r=0
+
+clean_exit() {
+    rc=$?
+    trap - INT TERM ABRT EXIT
+    if [ "x$TASK" != "x" ]; then
+        echo "Aborting from \"$TASK\" with the exit code $rc"
+    fi
+    mv /tmp/fc-build.log . || :
+    exit $rc
+}
+
+trap clean_exit INT TERM ABRT EXIT
 
 if [ x"$buildsys" == "xautotools" ]; then
     for i in "${enable[@]}"; do
@@ -91,21 +105,29 @@ if [ x"$buildsys" == "xautotools" ]; then
         fi
         . .gitlab-ci/${FC_DISTRO_NAME}-cross.sh
     fi
-    rm -rf "$BUILDDIR" "$PREFIX" || :
-    mkdir "$BUILDDIR" "$PREFIX"
+    if [ $clean_build -eq 1 ]; then
+        rm -rf "$BUILDDIR" "$PREFIX" || :
+        mkdir "$BUILDDIR" "$PREFIX"
+    fi
     cd "$BUILDDIR"
-    ../autogen.sh --prefix="$PREFIX" --disable-cache-build ${buildopt[*]} 2>&1 | tee /tmp/fc-build.log || r=$?
-    $MAKE V=1 2>&1 | tee -a /tmp/fc-build.log || r=$?
+    TASK="autogen.sh"
+    ../autogen.sh --prefix="$PREFIX" --disable-cache-build ${buildopt[*]} 2>&1 | tee /tmp/fc-build.log
+    TASK="make"
+    $MAKE V=1 2>&1 | tee -a /tmp/fc-build.log
     if [ $disable_check -eq 0 ]; then
-        $MAKE check V=1 2>&1 | tee -a /tmp/fc-build.log || r=$?
+        TASK="make check"
+        $MAKE check V=1 2>&1 | tee -a /tmp/fc-build.log
     fi
     if [ $enable_install -eq 1 ]; then
-        $MAKE install V=1 2>&1 | tee -a /tmp/fc-build.log || r=$?
+        TASK="make install"
+        $MAKE install V=1 2>&1 | tee -a /tmp/fc-build.log
     fi
     if [ $distcheck -eq 1 ]; then
-        $MAKE distcheck V=1 2>&1 | tee -a /tmp/fc-build.log || r=$?
+        TASK="make distcheck"
+        $MAKE distcheck V=1 2>&1 | tee -a /tmp/fc-build.log
     fi
 elif [ x"$buildsys" == "xmeson" ]; then
+    TASK="pip install"
     pip install meson
 #   tomli not required for Python >= 3.11
     pip install tomli
@@ -114,6 +136,7 @@ elif [ x"$buildsys" == "xmeson" ]; then
 
         # Update bindgen on Fontations builds to improve support for constants in fcint.h
         if [[ "$i" == "fontations" ]]; then
+            TASK="cargo install"
             cargo install bindgen-cli
             # Prepend the cargo bin directory to PATH
             if [[ -d "$HOME/.cargo/bin" ]]; then
@@ -124,6 +147,7 @@ elif [ x"$buildsys" == "xmeson" ]; then
             fi
         fi
     done
+    TASK=
     for i in "${disable[@]}"; do
         buildopt+=(-D$i=disabled)
     done
@@ -145,17 +169,25 @@ elif [ x"$buildsys" == "xmeson" ]; then
         . .gitlab-ci/$FC_DISTRO_NAME-cross.sh
     fi
     buildopt+=(--default-library=$type)
-    meson setup --prefix="$PREFIX" -Dnls=enabled -Dcache-build=disabled -Diconv=enabled ${buildopt[*]} "$BUILDDIR" 2>&1 | tee /tmp/fc-build.log || r=$?
-    meson compile -v -C "$BUILDDIR" 2>&1 | tee -a /tmp/fc-build.log || r=$?
+    if [ $clean_build -eq 1 ]; then
+        rm -rf $BUILDDIR || :
+    fi
+    TASK="meson setup"
+    meson setup --prefix="$PREFIX" -Dnls=enabled -Dcache-build=disabled -Diconv=enabled ${buildopt[*]} "$BUILDDIR" 2>&1 | tee /tmp/fc-build.log
+    TASK="meson compile"
+    meson compile -v -C "$BUILDDIR" 2>&1 | tee -a /tmp/fc-build.log
     if [ $disable_check -eq 0 ]; then
-        meson test -v -C "$BUILDDIR" 2>&1 | tee -a /tmp/fc-build.log || r=$?
+        TASK="meson test"
+        meson test -v -C "$BUILDDIR" 2>&1 | tee -a /tmp/fc-build.log
     fi
     if [ $enable_install -eq 1 ]; then
-        meson install -C "$BUILDDIR" 2>&1 | tee -a /tmp/fc-build.log || r=$?
+        TASK="meson install"
+        meson install -C "$BUILDDIR" 2>&1 | tee -a /tmp/fc-build.log
     fi
     if [ $distcheck -eq 1 ]; then
-        meson dist -C "$BUILDDIR" 2>&1 | tee -a /tmp/fc-build.log || r=$?
+        TASK="meson dist"
+        meson dist -C "$BUILDDIR" 2>&1 | tee -a /tmp/fc-build.log
     fi
 fi
-mv /tmp/fc-build.log . || :
+TASK=
 exit $r
